+++ index.html
@@ -2,11 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="stylesheet" href="./src/stylesheets/input.css" />
+
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-    <script src="https://cdn.tailwindcss.com"></script>
 
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
@@ -14,6 +12,8 @@
       href="https://fonts.googleapis.com/css2?family=League+Spartan:wght@500;700&display=swap"
       rel="stylesheet"
     />
+
+    <link rel="stylesheet" href="./src/stylesheets/input.css" />
     <title>Job Listing WebApp</title>
   </head>
   <body>
