--- src/App.css
@@ -1,32 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-:root {
-  --Desaturated-Dark-Cyan: hsl(180, 29%, 50%);
-  --Light-Grayish-Cyan: hsl(180, 52%, 96%);
-  --Light-Grayish-Cyan: hsl(180, 31%, 95%);
-  --Dark-Grayish-Cyan: hsl(180, 8%, 52%);
-  --Very-Dark-Grayish-Cyan: hsl(180, 14%, 20%);
-}
-
-.body-bg {
-  background: url("./images/bg-header-mobile.svg"), var(--Desaturated-Dark-Cyan);
-  background-repeat: no-repeat;
-}
-
-@media screen and (min-width: 700px) {
-  .body-bg {
-    background: url("./images/bg-header-desktop.svg"),
-      var(--Desaturated-Dark-Cyan);
-    background-repeat: no-repeat;
-  }
-}
-.list-value{
- padding:0.2rem 0.2rem;
- background-color: var(--Light-Grayish-Cyan);
- color: var(--Dark-Grayish-Cyan);
- text-transform: uppercase;
- font-size: 0.475rem;
- font-weight: 700;
-}
