+++ src/components/SearchBox.jsx
@@ -1,0 +1,11 @@
+export default function SearchBox() {
+  return (
+    <>
+      <div className="bg-[var(--Desaturated-Dark-Cyan)] bg-[url('/src/assets/images/bg-header-mobile.svg')] sm:bg-[url('/src/assets/images/bg-header-desktop.svg')] bg-cover bg-center bg-no-repeat min-h-[155px]"></div>
+
+      <div className="search-box flex mx-6  p-5 rounded-md shadow-[var(--card-shadow)] shadow-2xl">
+        Search Box
+      </div>
+    </>
+  );
+}
