--- src/components/searchComponent.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-import "../App.css";
-
-const searchComponent = () => {
-  return (
-    <div className="">
-      <form action="" className="py-4">
-        <input
-          type="text"
-          className=" outline-none bg-transparent py-4 bg-white break-words"
-        />
-      </form>
-    </div>
-  );
-};
-
-export default searchComponent;
