+++ src/stylesheets/input.css
@@ -8,6 +8,8 @@
   --Light-Grayish-Cyan2: hsl(180, 31%, 95%);
   --Dark-Grayish-Cyan: hsl(180, 8%, 52%);
   --Very-Dark-Grayish-Cyan: hsl(180, 14%, 20%);
-
   --card-shadow: hsla(180, 29%, 50%, 0.403);
+}
+*{
+  @apply font-['League_Spartan',sans-serif] ;
 }
